+++ lib/data/text_constants.dart
@@ -11,7 +11,7 @@ class TextConstants {
   static const Color defaultColor = Colors.black;
 
   /// color used if no color is specified
-  static const Color defaultBackgroundColor = Colors.white;
+  static const Color defaultBackgroundColor = Colors.transparent;
 
   /// link underline when no override was provided
   static const TextDecoration defaultLinkUnderline = TextDecoration.underline;