+++ lib/light_html_editor.dart
@@ -1,9 +1,12 @@
 export 'editor.dart' show LightHtmlRichTextEditor;
 export 'api/placeholder.dart' show RichTextPlaceholder;
+
 export 'ui/editor_elements/editor_code_renderer.dart' show LightHtmlRenderer;
 export 'ui/editor_elements/editor_code_textfield.dart'
     show LightHtmlEditorTextField;
+
 export 'data/editor_properties.dart' show EditorDecoration;
 export 'data/renderer_decoration.dart' show RendererStyle;
 export 'data/renderer_text_properties.dart' show RendererTextProperties;
 export 'api/v2/parser.dart' show Parser;
+export 'api/v3/parser.dart' show LightHtmlParserV3;